requirements.txt
Django==2.0.2
djangorestframework==3.7.7
django-configurations==2.0
psycopg2==2.7.4
dj-database-url==0.4.2
gunicorn==19.7.1
whitenoise==3.3.1